+++ lib/actions/datasource.js
@@ -9,7 +9,7 @@ class DataSourceActions {
     const workspace = this.getWorkspace();
     const datasource = this;
     const facet = workspace.facets(facetName);
-    fsUtility.writeDataSourceConfig(facet, datasource, function(err) {
+    this.write(facet, datasource, function(err) {
       if (err) return cb(err);
       facet.add(datasource);
       cb();
@@ -18,7 +18,7 @@ class DataSourceActions {
   refresh(facetName, cb) {
     const workspace = this.getWorkspace();
     const facet = workspace.facets(facetName);
-    fsUtility.readDataSource(facet, function(err, config) {
+    this.read(facet, function(err, config) {
       if (err) return cb(err);
       const datasources = facet.datasources();
       Object.keys(config).forEach(function(key) {
@@ -38,7 +38,7 @@ class DataSourceActions {
     const datasource = this;
     datasource.set(attrs);
     const facet = workspace.facets(facetName);
-    fsUtility.writeDataSourceConfig(facet, datasource, cb);
+    this.write(facet, datasource, cb);
   }
 }
