+++ lib/actions/facet.js
@@ -2,14 +2,14 @@
 
 const config = require('../config');
 const Facet = require('../datamodel/facet');
-const fsUtility = require('../util/file-utility');
 const ModelConfig = require('../datamodel/model-config');
 const mixin = require('../util/mixin');
 
 class FacetAction {
   refresh(cb) {
     const facet = this;
-    fsUtility.readModelConfig(facet, cb);
+    const modelConfig = new ModelConfig({}, 'id', {});
+    modelConfig.read(facet, cb);
   }
   create(facetDef, cb) {
     const workspace = this.getWorkspace();
@@ -21,7 +21,7 @@ class FacetAction {
       let modelConfig = new ModelConfig(workspace, config.name, config);
       facet.add(modelConfig);
     });
-    fsUtility.writeFacet(workspace, facet, function(err) {
+    this.write(function(err) {
       if (err) return cb(err);
       workspace.add(facet);
       cb();
