+++ lib/actions/middleware.js
@@ -10,11 +10,11 @@ class MiddlewareActions {
     const workspace = this.getWorkspace();
     const phase = workspace.phases(phaseName);
     phase.add(this);
-    fsUtility.writeMiddleware(workspace, [], cb);
+    this.write(workspace, [], cb);
   }
   refresh(cb) {
     const workspace = this.getWorkspace();
-    fsUtility.readMiddleware(workspace, cb);
+    this.read(workspace, cb);
   }
 }
 
@@ -23,7 +23,7 @@ mixin(Middleware.prototype, MiddlewareActions.prototype);
 class MiddlewarePhaseActions {
   refresh(cb) {
     const workspace = this.getWorkspace();
-    fsUtility.readMiddleware(workspace, cb);
+    this.read(workspace, cb);
   }
   create(phaseName, index, before, cb) {
     const workspace = this.getWorkspace();
@@ -32,7 +32,8 @@ class MiddlewarePhaseActions {
       return cb(new Error('invalid input phase'));
     index = findCorrectIndex(before, workspace);
     index = lastIndex(index, workspace);
-    fsUtility.writeMiddleware(workspace, phaseArr, function(err) {
+    const middleware = new Middleware({}, 'id', {});
+    middleware.write(workspace, phaseArr, function(err) {
       if (err) return cb(err);
       phaseArr.forEach(function(phase) {
         workspace.middlewarePhases.splice(index++, 0, phase);
