+++ lib/actions/model-config.js
@@ -10,7 +10,7 @@ class ModelConfigActions {
     const modelConfig = this;
     const facet = workspace.facet(facetName);
     const model = workspace.models(modelId);
-    fsUtility.writeModelConfig(facet, modelConfig, function(err) {
+    this.write(facet, modelConfig, function(err) {
       if (err) return cb(err);
       modelConfig.connect(facet, model);
       facet.add(modelConfig);
@@ -21,7 +21,7 @@ class ModelConfigActions {
     const workspace = this.getWorkspace();
     const modelConfig = this;
     modelConfig.set(attrs);
-    fsUtility.writeModelConfig(facet, modelConfig, cb);
+    this.write(facet, modelConfig, cb);
   }
 }
