+++ lib/actions/model-property.js
@@ -10,7 +10,7 @@ class ModelPropertyActions {
     const workspace = this.getWorkspace();
     const model = workspace.model(modelId);
     model.add(this);
-    fsUtility.writeModel(model, cb);
+    model.write(model, cb);
   }
 }
 
@@ -21,7 +21,7 @@ class ModelMethodActions {
     const workspace = this.getWorkspace();
     const model = workspace.model(modelId);
     model.add(this);
-    fsUtility.writeModel(model, cb);
+    model.write(model, cb);
   }
 }
