+++ lib/actions/model-relation.js
@@ -11,13 +11,13 @@ class ModelRelationActions {
     const toModel = workspace.model(toModelId);
     model.add(this);
     this.connect(model, toModel);
-    fsUtility.writeModel(model, cb);
+    model.write(model, cb);
   }
   delete(modelId, relationName, cb) {
     const workspace = this.getWorkspace();
     const model = workspace.models(modelId);
     model.removeRelation(relationName);
-    fsUtility.writeModel(model, cb);
+    model.write(model, cb);
   }
 }
