+++ lib/actions/model.js
@@ -11,7 +11,7 @@ class ModelActions {
   create(cb) {
     const workspace = this.getWorkspace();
     const self = this;
-    fsUtility.writeModel(self, function(err) {
+    this.write(self, function(err) {
       if (err) return cb(err);
       workspace.add(self);
       cb();
@@ -20,7 +20,7 @@ class ModelActions {
   refresh(cb) {
     const workspace = this.getWorkspace();
     const model = this;
-    fsUtility.readModel(
+    this.read(
       model.getFacetName(),
       model.getName(),
       workspace,
@@ -35,14 +35,14 @@ class ModelActions {
     const workspace = this.getWorkspace();
     const model = this;
     model.set(attrs);
-    fsUtility.writeModel(model, cb);
+    this.write(model, cb);
   }
   delete(cb) {
     const workspace = this.getWorkspace();
     const model = this;
     const err = model.remove();
     if (err) return cb(err);
-    fsUtility.removeModel(model, cb);
+    this.unlink(model, cb);
   }
 }
