+++ lib/actions/package-definition.js
@@ -14,7 +14,7 @@ const mixin = require('../util/mixin');
 class PackageDefinitionActions {
   create(cb) {
     const workspace = this.getWorkspace();
-    fsUtility.writePackageDefinition(this, cb);
+    this.write(this, cb);
   }
 }
