+++ lib/datamodel/datasource.js
@@ -1,5 +1,7 @@
 'use strict';
 const Entity = require('./entity');
+const fs = require('fs-extra');
+const path = require('path');
 
 /**
  * @class DataSource
@@ -13,6 +15,25 @@ class DataSource extends Entity {
   }
   getDefinition() {
     return this._content;
+  }
+  read(facet, cb) {
+    const filePath = facet.getDataSourceConfigFilePath();
+    fs.readJson(filePath, function(err, data) {
+      if (err) return err;
+      cb(null, data);
+    });
+  }
+  write(facet, datasource, cb) {
+    const configData = facet.datasources().map({json: true});
+    configData[datasource._name] = datasource.getContents();
+    const filePath = facet.getDataSourceConfigFilePath();
+    fs.mkdirp(path.dirname(filePath), function(err) {
+      if (err) return cb(err);
+      fs.writeJson(filePath, configData, function(err) {
+        if (err) return cb(err);
+        cb(null, configData);
+      });
+    });
   }
 };
