+++ lib/datamodel/entity.js
@@ -1,6 +1,7 @@
 'use strict';
 const clone = require('lodash').clone;
 const Composite = require('./graph').Composite;
+const fs = require('fs-extra');
 
 /**
  * @class Entity
