+++ lib/datamodel/facet-config.js
@@ -1,5 +1,6 @@
 'use strict';
 const Entity = require('./entity');
+const fs = require('fs-extra');
 
 /**
  * @class FacetConfig
@@ -10,5 +11,12 @@ class FacetConfig extends Entity {
   constructor(Workspace, name, data, options) {
     super(Workspace, 'FacetConfig', name, data);
   }
+
+  write(facet, cb) {
+    const facetConfigFile = facet.getConfigPath();
+    const facetConfig = facet.getConfig();
+    fs.writeJson(facetConfigFile, facetConfig, cb);
+  }
+
 }
 module.exports = FacetConfig;
