+++ lib/datamodel/facet.js
@@ -1,10 +1,13 @@
 'use strict';
+
+const async = require('async');
 const config = require('../config.json');
 const Entity = require('./entity');
 const path = require('path');
 const ModelConfig = require('./model-config');
 const FacetConfig = require('./facet-config');
 const Datasource = require('./datasource');
+const fs = require('fs-extra');
 
 /**
  * @class Facet
@@ -69,6 +72,23 @@ class Facet extends Entity {
     const filePath = path.join(workspace.directory, this._name,
       config.DataSourceConfigFile);
     return filePath;
+  }
+  write(cb) {
+    const facet = this;
+    const facetFolder = facet.getPath();
+    const tasks = [];
+    tasks.push(function(next) {
+      fs.mkdirp(facetFolder, next);
+    });
+    tasks.push(function(next) {
+      const facetConfig = new FacetConfig({}, 'id', {});
+      facetConfig.write(facet, next);
+    });
+    tasks.push(function(next) {
+      const modelConfig = new ModelConfig({}, 'id', {});
+      modelConfig.write(facet, null, next);
+    });
+    async.series(tasks, cb);
   }
 };
