+++ lib/datamodel/middleware-phase.js
@@ -1,6 +1,7 @@
 'use strict';
 const Entity = require('./entity');
 const Middleware = require('./middleware');
+const fs = require('fs-extra');
 
 /**
  * @class MiddlewarePhase
