+++ lib/datamodel/middleware.js
@@ -1,5 +1,7 @@
 'use strict';
 const Entity = require('./entity');
+const fs = require('fs-extra');
+const path = require('path');
 
 /**
  * @class Middleware
@@ -18,6 +20,28 @@ class Middleware extends Entity {
   }
   setConfig(config) {
     this._content = config;
+  }
+  read(workspace, cb) {
+    const filePath = workspace.getMiddlewareFilePath();
+    fs.readJson(filePath, function(err, data) {
+      if (err) return cb(err);
+      workspace.setMiddlewareConfig(data);
+      cb(null, data);
+    });
+  }
+  write(workspace, phaseArr, cb) {
+    const data = workspace.getMiddlewareConfig();
+    phaseArr.forEach(function(phaseName) {
+      data[phaseName] = {};
+    });
+    const file = workspace.getMiddlewareFilePath();
+    fs.mkdirp(path.dirname(file), function(err) {
+      if (err) return cb(err);
+      fs.writeJson(file, data, function(err) {
+        if (err) return cb(err);
+        cb();
+      });
+    });
   }
 };
