+++ lib/datamodel/model-method.js
@@ -1,5 +1,6 @@
 'use strict';
 const Entity = require('./entity');
+const fs = require('fs-extra');
 
 /**
  * @class ModelMethod
