+++ lib/datamodel/model-relationship.js
@@ -1,5 +1,6 @@
 'use strict';
 const Relation = require('../datamodel/relation');
+const fs = require('fs-extra');
 
 /**
  * @class ModelRelation
