+++ lib/datamodel/model.js
@@ -7,6 +7,7 @@ const path = require('path');
 const ModelRelation = require('./model-relationship');
 const ModelMethod = require('./model-method');
 const ModelProperty = require('./model-property');
+const fs = require('fs-extra');
 
 /**
  * @class Model
@@ -69,6 +70,23 @@ class Model extends Entity {
       relation.remove();
       model.relations().remove(relation);
     }
+  }
+  read(facetName, modelName, workspace, cb) {
+    const filePath = workspace.getModelDefinitionPath(facetName, modelName);
+    fs.readJson(filePath, cb);
+  }
+  write(model, cb) {
+    const filePath = model.getFilePath();
+    const data = model.getDefinition();
+    const dir = path.dirname(filePath);
+    fs.mkdirp(dir, function(err) {
+      if (err) return cb(err);
+      fs.writeJson(filePath, data, cb);
+    });
+  }
+  unlink(model, cb) {
+    const filePath = model.getFilePath();
+    fs.unlink(filePath, cb);
   }
 };
