+++ lib/datamodel/package-definition.js
@@ -1,6 +1,7 @@
 'use strict';
 const Entity = require('./entity');
 const path = require('path');
+const fs = require('fs-extra');
 
 /**
  * @class PackageDefinition
@@ -17,6 +18,11 @@ class PackageDefinition extends Entity {
   }
   getDefinition() {
     return this._content;
+  }
+  write(packageDef, cb) {
+    const filePath = packageDef.getFilePath();
+    const data = packageDef.getDefinition();
+    fs.writeJson(filePath, data, cb);
   }
 };
