+++ lib/datamodel/relation.js
@@ -1,6 +1,7 @@
 'use strict';
 const Entity = require('./entity');
 const Edge = require('./graph/edge');
+const fs = require('fs-extra');
 
 /**
  * @class Relation
