+++ lib/util/file-utility.js
@@ -13,141 +13,7 @@ const path = require('path');
 module.exports = {
   getConfigFiles: getConfigFiles,
   readFile: readFile,
-  readModel: readModel,
-  removeModel: removeModel,
-  readModelConfig: readModelConfig,
-  readDataSource: readDataSource,
-  readMiddleware: readMiddleware,
-  writeDataSourceConfig: writeDataSourceConfig,
-  writeFacet: writeFacet,
-  writeFacetConfig: writeFacetConfig,
-  writeModel: writeModel,
-  writeMiddleware: writeMiddleware,
-  writeModelConfig: writeModelConfig,
-  writePackageDefinition: writePackageDefinition,
 };
-
-function writeFacet(workspace, facet, cb) {
-  const facetFolder = facet.getPath();
-  const tasks = [];
-  tasks.push(function(next) {
-    fs.mkdirp(facetFolder, next);
-  });
-  tasks.push(function(next) {
-    writeFacetConfig(facet, next);
-  });
-  tasks.push(function(next) {
-    writeModelConfig(facet, null, next);
-  });
-  async.series(tasks, cb);
-}
-
-function writeFacetConfig(facet, cb) {
-  const facetConfigFile = facet.getConfigPath();
-  const facetConfig = facet.getConfig();
-  fs.writeJson(facetConfigFile, facetConfig, cb);
-}
-
-function writeModelConfig(facet, modelConfig, cb) {
-  let data = {};
-  data._meta = config.modelsMetadata;
-  if (modelConfig)
-    data[modelConfig._name] =
-      modelConfig.getContents({filter: ['id', 'modelId']});
-  const modelConfigData =
-    facet.modelconfig().map({json: true, filter: 'modelId'});
-  data = Object.assign({}, data, modelConfigData);
-  const filePath = facet.getModelConfigPath();
-  fs.writeJson(filePath, data, cb);
-}
-
-function writeModel(model, cb) {
-  const filePath = model.getFilePath();
-  const data = model.getDefinition();
-  const dir = path.dirname(filePath);
-  fs.mkdirp(dir, function(err) {
-    if (err) return cb(err);
-    fs.writeJson(filePath, data, cb);
-  });
-}
-
-function removeModel(model, cb) {
-  const filePath = model.getFilePath();
-  fs.unlink(filePath, cb);
-}
-
-function writeDataSourceConfig(facet, datasource, cb) {
-  const configData = facet.datasources().map({json: true});
-  configData[datasource._name] = datasource.getContents();
-  const filePath = facet.getDataSourceConfigFilePath();
-  fs.mkdirp(path.dirname(filePath), function(err) {
-    if (err) return cb(err);
-    fs.writeJson(filePath, configData, function(err) {
-      if (err) return cb(err);
-      cb(null, configData);
-    });
-  });
-}
-
-function writeMiddleware(workspace, phaseArr, cb) {
-  const data = workspace.getMiddlewareConfig();
-  phaseArr.forEach(function(phaseName) {
-    data[phaseName] = {};
-  });
-  const file = workspace.getMiddlewareFilePath();
-  fs.mkdirp(path.dirname(file), function(err) {
-    if (err) return cb(err);
-    fs.writeJson(file, data, function(err) {
-      if (err) return cb(err);
-      cb();
-    });
-  });
-}
-
-function writePackageDefinition(packageDef, cb) {
-  const filePath = packageDef.getFilePath();
-  const data = packageDef.getDefinition();
-  fs.writeJson(filePath, data, cb);
-}
-
-function readModelConfig(facet, cb) {
-  const filePath = facet.getModelConfigPath();
-  fs.readJson(filePath, function(err, config) {
-    if (err) return err;
-    Object.keys(config).forEach(function(key) {
-      if (key === '_meta') return;
-      let modelConfig = facet.modelconfig(key);
-      if (modelConfig) {
-        modelConfig._content = config[key];
-      } else {
-        facet.setModelConfig(key, config[key]);
-      }
-    });
-    cb(null, config);
-  });
-}
-
-function readModel(facetName, modelName, workspace, cb) {
-  const filePath = workspace.getModelDefinitionPath(facetName, modelName);
-  fs.readJson(filePath, cb);
-}
-
-function readDataSource(facet, cb) {
-  const filePath = facet.getDataSourceConfigFilePath();
-  fs.readJson(filePath, function(err, data) {
-    if (err) return err;
-    cb(null, data);
-  });
-}
-
-function readMiddleware(workspace, cb) {
-  const filePath = workspace.getMiddlewareFilePath();
-  fs.readJson(filePath, function(err, data) {
-    if (err) return cb(err);
-    workspace.setMiddlewareConfig(data);
-    cb(null, data);
-  });
-}
 
 function readFile(filePath, cb) {
   fs.readJson(filePath, cb);
